refactor: split compilation mod
